move severity tests to separate file
